separate InitSessionBuilder out
