+++ Branch-SDK-Automation-TestBed/src/main/java/io/branch/saas/sdk/testbed/activities/ReadDeepLinkActivity.java
@@ -17,6 +17,7 @@ import org.json.JSONObject;
 
 import io.branch.referral.Branch;
 import io.branch.referral.BranchError;
+import io.branch.referral.InitSessionBuilder;
 import io.branch.saas.sdk.testbed.Common;
 import io.branch.saas.sdk.testbed.Constants;
 import io.branch.saas.sdk.testbed.R;
@@ -87,7 +88,7 @@ public class ReadDeepLinkActivity extends AppCompatActivity implements AdvancedW
             return;
         }
         Common.getInstance().clearLog();
-        Branch.InitSessionBuilder initSessionBuilder = Branch.sessionBuilder(this).withCallback(new Branch.BranchReferralInitListener() {
+        InitSessionBuilder initSessionBuilder = Branch.sessionBuilder(this).withCallback(new Branch.BranchReferralInitListener() {
             @Override
             public void onInitFinished(JSONObject referringParams, BranchError error) {
                 if (error == null) {
